Latest version of autotag
Multiple elements per line are now supported. Self closing tags will not trigger the plugin. Make sure that you don't try and skip back a few tags to enter a tag you might have forgotten, in that case you'll have to type in your own closing tags.

If you're opening a tag that already has a closing tag in the document and you want more than one of that particular tag just enter an extra chevron ">" to trigger the autocomplete